pyproject.toml
[project]
name = "llm-data-selection-bert"
version = "0.1.0"
requires-python = ">=3.13,<3.14"
description = "Data subset selection experiments on SST-2 with ModernBERT"
dependencies = [
  "numpy>=2.3.4",
  "pandas>=2.3.3",
  "transformers>=4.57.1",
  "datasets>=4.2.0",
  "hydra-core>=1.3.2",
  "wandb>=0.22.2",
  "scikit-learn>=1.7.2",
  "evaluate>=0.4.6",
  "accelerate>=1.10.1",
  "matplotlib>=3.10.7",
  "umap-learn>=0.5.9",
  "faiss-cpu>=1.12.0",
  "python-dotenv>=1.1.1",
  "pyyaml>=6.0.3",
  "snakemake>=9.13.3",
  "textstat>=0.7.10",
  "joblib>=1.5.2",
  "sentencepiece>=0.2.1",
  "lexicalrichness>=0.5.1",
  "evaluate>=0.4.6"
]

[project.optional-dependencies]
cpu = [
  "torch>=2.6,<2.7"
]

gpu = [
  "torch>=2.6,<2.7",
  "nvidia-cublas-cu12",
  "nvidia-cudnn-cu12"
]

[tool.uv]
conflicts = [
  [ { extra = "cpu" }, { extra = "gpu" } ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu",   extra = "cpu" },
  { index = "pytorch-cu124", extra = "gpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url  = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url  = "https://download.pytorch.org/whl/cu124"
explicit = true
